Improve temp-table query
The method that checks for the use of temporary tables cleans its
statement before it runs the query.  This leaves the statement in a
"dirty" state between checks, which locks the tables_used table.  This
change cleans the statement after the query is run so that no locks
are held.

The unit test has been updated.

Flag: EXEMPT bugfix
Bug: 342497314
Test: atest
 * CtsDatabaseTestCases
 * FrameworksCoreTests:android.database
Change-Id: I54032d8c360b123596ca73e6c294dbb285e05361
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>